backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anthropic==0.75.0
anyio==4.12.1
APScheduler==3.10.4
attrs==23.1.0
bcrypt==3.2.0
blinker==1.9.0
cachetools==6.2.4
certifi==2024.2.2
cffi==2.0.0
charset-normalizer==3.4.4
click==8.1.7
colorama==0.4.6
colorhash==2.1.0
configparser==7.2.0
contourpy==1.3.2
cryptography==46.0.5
cycler==0.12.1
dataclasses-json==0.6.7
distro==1.9.0
dnspython==2.8.0
docstring_parser==0.17.0
ecdsa==0.19.1
email-validator==2.3.0
fastapi==0.128.0
file-operations==1.0.5
filelock==3.20.2
filetype==1.2.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-MonitoringDashboard==3.2.1
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
fonttools==4.58.4
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.29.0
google-api-python-client==2.187.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.57.0
google-generativeai==0.8.6
googleapis-common-protos==1.72.0
gpt4all==2.8.2
greenlet==3.2.3
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
httpx-sse==0.4.3
huggingface-hub==0.36.0
idna==3.11
importlib_metadata==7.1.0
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.12.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.21.1
jsonschema-specifications==2025.4.1
kiwisolver==1.4.5
kneed==0.8.5
langchain==1.2.3
langchain-anthropic==1.3.1
langchain-classic==1.0.1
langchain-community==0.4.1
langchain-core==1.2.6
langchain-google-genai==4.1.3
langchain-huggingface==1.2.0
langchain-text-splitters==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.6.2
MarkupSafe==2.1.5
marshmallow==3.26.2
matplotlib==3.7.1
more-itertools==10.2.0
motor==3.7.1
mpmath==1.3.0
multidict==6.7.0
mypy_extensions==1.1.0
narwhals==2.15.0
numpy==2.4.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
pandas==1.5.3
passlib==1.7.4
pillow==11.2.1
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
psutil==7.0.0
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pydeck==0.9.1
pymongo==4.16.0
pyparsing==3.1.2
pyrsistent==0.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-jose==3.5.0
python-multipart==0.0.22
pytz==2024.1
PyYAML==6.0.1
referencing==0.36.2
regex==2025.11.3
requests==2.32.5
requests-toolbelt==1.0.0
rpds-py==0.25.1
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.2.2
scipy==1.15.3
six==1.16.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.50.0
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
transformers==4.57.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==2.0.0
uritemplate==4.2.0
urllib3==2.6.3
uuid_utils==0.13.0
uvicorn==0.40.0
watchdog==6.0.0
websockets==15.0.1
Werkzeug==2.3.8
WTForms==3.2.1
xgboost==1.7.4
xxhash==3.6.0
yarl==1.22.0
zipp==3.18.1
zstandard==0.25.0
